Code for all the sensors (IR, ODS, Color), code for Servo motors
This checkin has the code for all the sensors and the servo motor.
The key is the following.
ODS sensor does not detect distance as the name suggets. It detects
light. I recommend using it under the robot for line detection.
The color sensors and the servo work. All of them have init method, a
test method and variables to store data.